src/discord/bot/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
coloredlogs==14.0
discord==1.0.1
discord.py==1.3.2
humanfriendly==8.1
idna==2.9
multidict==4.7.5
python-dotenv==0.12.0
websockets==8.1
yarl==1.4.2
